Removed truncating of enum vals to int in ReverseLookup.
Some implementations (e.g. C++98) won't support 64-bit enum values,
but there is no reason to silently truncate them.

Tested: on Linux.

+++ include/flatbuffers/idl.h
@@ -317,7 +317,7 @@ struct EnumVal {
 struct EnumDef : public Definition {
   EnumDef() : is_union(false), uses_type_aliases(false) {}
 
-  EnumVal *ReverseLookup(int enum_idx, bool skip_union_default = true) {
+  EnumVal *ReverseLookup(int64_t enum_idx, bool skip_union_default = true) {
     for (auto it = vals.vec.begin() +
                    static_cast<int>(is_union && skip_union_default);
          it != vals.vec.end(); ++it) {
